+++ src/components/LinkChooser/SearchResultsList.tsx
@@ -163,7 +163,7 @@ const SearchResultOption: FC<SearchResultOptionProps> = ({ item, state, keyItemR
 const EmptyResults: FC<{ prompt: string; icon: string }> = ({ prompt, icon }) => (
     <div
         data-test-id="link-chooser-empty-results"
-        className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[300px] tw-p-3"
+        className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[220px] tw-p-3"
     >
         <img className="tw-w-[75px] tw-mb-5" src={icon} alt="Icon" />
         <p className="tw-text-black-60 tw-text-center">{prompt}</p>
@@ -173,7 +173,7 @@ const EmptyResults: FC<{ prompt: string; icon: string }> = ({ prompt, icon }) =>
 const FetchingError: FC<{ error?: string }> = ({ error = "An error occurred while fetching the results" }) => (
     <div
         data-test-id="link-chooser-error"
-        className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[300px] tw-p-3"
+        className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[220px] tw-p-3"
     >
         <img className="tw-w-[75px] tw-mb-5" src={NoResultsIcon} alt="Error" />
         <p className="tw-text-black-60 tw-text-center">{error}</p>
@@ -183,7 +183,7 @@ const FetchingError: FC<{ error?: string }> = ({ error = "An error occurred whil
 const FetchingAnimation: FC = () => (
     <div
         data-test-id="link-chooser-loader"
-        className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[300px] tw-p-3"
+        className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[220px] tw-p-3"
     >
         <img className="tw-w-[50px]" src={FetchingIcon} alt="Fetching" />
     </div>
@@ -191,11 +191,11 @@ const FetchingAnimation: FC = () => (
 
 const ImageMenuItem: FC<ImageMenuItemProps> = ({ title, subtitle, preview }) => (
     <div className="tw-flex tw-px-5 tw-py-1.5 tw-cursor-pointer">
-        <div className="tw-flex tw-flex-shrink-0 tw-w-[75px] tw-h-[75px] tw-max-w-xs tw-bg-black-0 tw-border-black-10 tw-border tw-rounded">
+        <div className="tw-flex tw-flex-shrink-0 tw-w-[75px] tw-h-[75px] tw-max-w-xs tw-bg-black-0 tw-rounded">
             {preview && <img className="tw-w-full tw-object-contain" src={preview} alt={title as string} />}
         </div>
         <div className="tw-flex tw-flex-col tw-justify-center tw-ml-4">
-            <p className="tw-block tw-text-lg tw-leading-tight tw-hover:underline tw-text-black-80">{title}</p>
+            <p className="tw-block tw-text-md tw-leading-tight tw-hover:underline tw-text-black-80">{title}</p>
             {subtitle && <p className="tw-text-[12px] tw-text-black-80">{subtitle}</p>}
         </div>
     </div>
